Create auth package
